+++ Test/Braintree/OAuthTestHelper.php
@@ -7,9 +7,10 @@ class OAuthTestHelper
 {
     public static function createGrant($gateway, $params)
     {
-        $http = new Braintree\HttpOAuth($gateway->config);
-        $response = $http->post('/oauth_testing/grants', $params);
-        return $response['code'];
+        $http = new Braintree\Http($gateway->config);
+        $http->useClientCredentials();
+        $response = $http->post('/oauth_testing/grants', array('grant' => $params));
+        return $response['grant']['code'];
     }
 
     public static function createCredentials($params)