+++ Test/Integration/MerchantTest.php
@@ -18,6 +18,7 @@ class MerchantTest extends Setup
         $result = $gateway->merchant()->create(array(
             'email' => 'name@email.com',
             'countryCodeAlpha3' => 'USA',
+            'paymentMethods' => ['credit_card', 'paypal']
         ));
 
         $this->assertEquals(true, $result->success);
@@ -29,7 +30,7 @@ class MerchantTest extends Setup
 
     /**
     * @expectedException Exception\Configuration
-    * @expectedExceptionMessage clientId needs to be set.
+    * @expectedExceptionMessage clientId needs to be passed to Braintree\Gateway.
     */
     public function testAssertsHasCredentials()
     {
@@ -40,5 +41,22 @@ class MerchantTest extends Setup
             'email' => 'name@email.com',
             'countryCodeAlpha3' => 'USA',
         ));
+    }
+
+    public function testBadPaymentMethods()
+    {
+        $gateway = new Braintree\Gateway(array(
+            'clientId' => 'client_id$development$integration_client_id',
+            'clientSecret' => 'client_secret$development$integration_client_secret',
+        ));
+        $result = $gateway->merchant()->create(array(
+            'email' => 'name@email.com',
+            'countryCodeAlpha3' => 'USA',
+            'paymentMethods' => ['fake_money'],
+        ));
+
+        $this->assertEquals(false, $result->success);
+        $errors = $result->errors->forKey('merchant')->onAttribute('paymentMethods');
+        $this->assertEquals(Braintree\Error\Codes::MERCHANT_ACCOUNT_PAYMENT_METHODS_ARE_INVALID, $errors[0]->code);
     }
 }