+++ Test/Integration/OAuthTest.php
@@ -10,8 +10,8 @@ use Braintree;
 class OAuthTest extends Setup
 {
     /**
-    * @expectedException Braintree\Exception\Configuration
-    * @expectedExceptionMessage clientSecret needs to be set.
+    * @expectedException Braintree\Exception_Configuration
+    * @expectedExceptionMessage clientSecret needs to be passed to Braintree\Gateway.
     */
     public function testAssertsHasCredentials()
     {
@@ -39,13 +39,77 @@ class OAuthTest extends Setup
         ));
 
         $this->assertEquals(true, $result->success);
+        $credentials = $result->credentials;
+        $this->assertNotNull($credentials->accessToken);
+        $this->assertNotNull($credentials->refreshToken);
+        $this->assertEquals('bearer', $credentials->tokenType);
+        $this->assertNotNull($credentials->expiresAt);
+    }
+
+    public function testCreateTokenFromCodeWithMixedCredentials()
+    {
+        $gateway = new Braintree\Gateway(array(
+            'clientId' => 'client_id$development$integration_client_id',
+            'clientSecret' => 'client_secret$development$integration_client_secret',
+            'accessToken' => 'access_token$development$integration_merchant_id$f9ac33b3dd',
+        ));
+        $code = Test\Braintree\OAuthTestHelper::createGrant($gateway, array(
+            'merchant_public_id' => 'integration_merchant_id',
+            'scope' => 'read_write'
+        ));
+        $result = $gateway->oauth()->createTokenFromCode(array(
+            'code' => $code,
+            'scope' => 'read_write',
+        ));
+
+        $this->assertEquals(true, $result->success);
+        $credentials = $result->credentials;
+        $this->assertNotNull($credentials->accessToken);
+        $this->assertNotNull($credentials->refreshToken);
+        $this->assertEquals('bearer', $credentials->tokenType);
+        $this->assertNotNull($credentials->expiresAt);
+    }
+
+    public function testCreateTokenFromCode_JsonAPI()
+    {
+        $gateway = new Braintree\Gateway(array(
+            'clientId' => 'client_id$development$integration_client_id',
+            'clientSecret' => 'client_secret$development$integration_client_secret'
+        ));
+        $code = Test\Braintree\OAuthTestHelper::createGrant($gateway, array(
+            'merchant_public_id' => 'integration_merchant_id',
+            'scope' => 'read_write'
+        ));
+        $result = $gateway->oauth()->createTokenFromCode(array(
+            'code' => $code,
+            'scope' => 'read_write',
+        ));
+
+        $this->assertEquals(true, $result->success);
         $this->assertNotNull($result->accessToken);
         $this->assertNotNull($result->refreshToken);
         $this->assertEquals('bearer', $result->tokenType);
         $this->assertNotNull($result->expiresAt);
     }
 
-    public function testCreateTokenFromCodeFail()
+    public function testCreateTokenFromCode_ValidationErrorTest()
+    {
+        $gateway = new Braintree\Gateway(array(
+            'clientId' => 'client_id$development$integration_client_id',
+            'clientSecret' => 'client_secret$development$integration_client_secret'
+        ));
+        $result = $gateway->oauth()->createTokenFromCode(array(
+            'code' => 'bad_code',
+            'scope' => 'read_write',
+        ));
+
+        $this->assertEquals(false, $result->success);
+        $errors = $result->errors->forKey('credentials')->onAttribute('code');
+        $this->assertEquals(Braintree\Error\Codes::OAUTH_INVALID_GRANT, $errors[0]->code);
+        $this->assertEquals(1, preg_match('/Invalid grant: code not found/', $result->message));
+    }
+
+    public function testCreateTokenFromCode_OldError()
     {
         $gateway = new Braintree\Gateway(array(
             'clientId' => 'client_id$development$integration_client_id',
@@ -74,7 +138,7 @@ class OAuthTest extends Setup
         $refreshToken = $gateway->oauth()->createTokenFromCode(array(
             'code' => $code,
             'scope' => 'read_write',
-        ))->refreshToken;
+        ))->credentials->refreshToken;
 
         $result = $gateway->oauth()->createTokenFromRefreshToken(array(
             'refreshToken' => $refreshToken,
@@ -82,10 +146,11 @@ class OAuthTest extends Setup
         ));
 
         $this->assertEquals(true, $result->success);
-        $this->assertNotNull($result->accessToken);
-        $this->assertNotNull($result->refreshToken);
-        $this->assertEquals('bearer', $result->tokenType);
-        $this->assertNotNull($result->expiresAt);
+        $credentials = $result->credentials;
+        $this->assertNotNull($credentials->accessToken);
+        $this->assertNotNull($credentials->refreshToken);
+        $this->assertEquals('bearer', $credentials->tokenType);
+        $this->assertNotNull($credentials->expiresAt);
     }
 
 
@@ -132,6 +197,7 @@ class OAuthTest extends Setup
                 'currency' => 'USD',
                 'website' => 'http://example.com',
             ),
+            'paymentMethods' => ['credit_card'],
         ));
 
         $components = parse_url($url);
@@ -175,6 +241,10 @@ class OAuthTest extends Setup
         $this->assertEquals(7, $query['business']['fulfillment_completed_in']);
         $this->assertEquals('USD', $query['business']['currency']);
         $this->assertEquals('http://example.com', $query['business']['website']);
+
+        $this->assertCount(1, $query['payment_methods']);
+        $this->assertEquals('credit_card', $query['payment_methods'][0]);
+
         $this->assertEquals(64, strlen($query['signature']));
         $this->assertEquals('SHA256', $query['algorithm']);
     }
@@ -194,5 +264,21 @@ class OAuthTest extends Setup
         $this->assertArrayNotHasKey('merchant_id', $query);
         $this->assertArrayNotHasKey('redirect_uri', $query);
         $this->assertArrayNotHasKey('scope', $query);
+    }
+
+    public function testBuildConnectUrlWithPaymentMethods()
+    {
+        $gateway = new Braintree\Gateway(array(
+            'clientId' => 'client_id$development$integration_client_id',
+            'clientSecret' => 'client_secret$development$integration_client_secret'
+        ));
+        $url = $gateway->oauth()->connectUrl(array(
+            'paymentMethods' => array('credit_card', 'paypal')
+        ));
+
+        $queryString = parse_url($url)['query'];
+        parse_str($queryString, $query);
+
+        $this->assertEquals(array('credit_card', 'paypal'), $query['payment_methods']);
     }
 }