+++ src/Braintree/Http.php
@@ -7,9 +7,10 @@ namespace Braintree;
  *
  * @copyright  2014 Braintree, a division of PayPal, Inc.
  */
-class Http extends HttpBase
+class Http
 {
     protected $_config;
+    private $_useClientCredentials = false;
 
     public function __construct($config)
     {
@@ -19,47 +20,43 @@ class Http extends HttpBase
     public function delete($path)
     {
         $response = $this->_doRequest('DELETE', $path);
-
-        if ($response['status'] !== 200) {
+        if ($response['status'] === 200) {
+            return true;
+        } else {
             Util::throwStatusCodeException($response['status']);
         }
-
-        return true;
     }
 
     public function get($path)
     {
         $response = $this->_doRequest('GET', $path);
-
-        if ($response['status'] !== 200) {
+        if ($response['status'] === 200) {
+            return Xml::buildArrayFromXml($response['body']);
+        } else {
             Util::throwStatusCodeException($response['status']);
         }
-
-        return Xml::buildArrayFromXml($response['body']);
     }
 
     public function post($path, $params = null)
     {
         $response = $this->_doRequest('POST', $path, $this->_buildXml($params));
         $responseCode = $response['status'];
-
-        if (!in_array($responseCode, array(200, 201, 422), true)) {
+        if ($responseCode === 200 || $responseCode === 201 || $responseCode === 422 || $responseCode === 400) {
+            return Xml::buildArrayFromXml($response['body']);
+        } else {
             Util::throwStatusCodeException($responseCode);
         }
-
-        return Xml::buildArrayFromXml($response['body']);
     }
 
     public function put($path, $params = null)
     {
         $response = $this->_doRequest('PUT', $path, $this->_buildXml($params));
         $responseCode = $response['status'];
-
-        if (!in_array($responseCode, array(200, 201, 422), true)) {
+        if ($responseCode === 200 || $responseCode === 201 || $responseCode === 422 || $responseCode === 400) {
+            return Xml::buildArrayFromXml($response['body']);
+        } else {
             Util::throwStatusCodeException($responseCode);
         }
-
-        return Xml::buildArrayFromXml($response['body']);
     }
 
     private function _buildXml($params)
@@ -67,7 +64,7 @@ class Http extends HttpBase
         return empty($params) ? null : Xml::buildXmlFromArray($params);
     }
 
-    protected function _getHeaders()
+    private function _getHeaders()
     {
         return array(
             'Accept: application/xml',
@@ -75,16 +72,16 @@ class Http extends HttpBase
         );
     }
 
-    protected function _getAuthorization()
+    private function _getAuthorization()
     {
-        if ($this->_config->isAccessToken()) {
-            return array(
-                'token' => $this->_config->getAccessToken(),
-            );
-        } elseif ($this->_config->isClientCredentials()) {
+        if ($this->_useClientCredentials) {
             return array(
                 'user' => $this->_config->getClientId(),
                 'password' => $this->_config->getClientSecret(),
+            );
+        } elseif ($this->_config->isAccessToken()) {
+            return array(
+                'token' => $this->_config->getAccessToken(),
             );
         } else {
             return array(
@@ -92,5 +89,59 @@ class Http extends HttpBase
                 'password' => $this->_config->getPrivateKey(),
             );
         }
+    }
+
+    public function useClientCredentials()
+    {
+        $this->_useClientCredentials = true;
+    }
+
+    private function _doRequest($httpVerb, $path, $requestBody = null)
+    {
+        return $this->_doUrlRequest($httpVerb, $this->_config->baseUrl() . $path, $requestBody);
+    }
+
+    public function _doUrlRequest($httpVerb, $url, $requestBody = null)
+    {
+        $curl = curl_init();
+        curl_setopt($curl, CURLOPT_TIMEOUT, 60);
+        curl_setopt($curl, CURLOPT_CUSTOMREQUEST, $httpVerb);
+        curl_setopt($curl, CURLOPT_URL, $url);
+        curl_setopt($curl, CURLOPT_ENCODING, 'gzip');
+
+        $headers = $this->_getHeaders($curl);
+        $headers[] = 'User-Agent: Braintree PHP Library ' . Version::get();
+        $headers[] = 'X-ApiVersion: ' . Configuration::API_VERSION;
+
+        $authorization = $this->_getAuthorization();
+        if (isset($authorization['user'])) {
+            curl_setopt($curl, CURLOPT_HTTPAUTH, CURLAUTH_BASIC);
+            curl_setopt($curl, CURLOPT_USERPWD, $authorization['user'] . ':' . $authorization['password']);
+        } elseif (isset($authorization['token'])) {
+            $headers[] = 'Authorization: Bearer ' . $authorization['token'];
+        }
+        curl_setopt($curl, CURLOPT_HTTPHEADER, $headers);
+
+        // curl_setopt($curl, CURLOPT_VERBOSE, true);
+        if ($this->_config->sslOn()) {
+            curl_setopt($curl, CURLOPT_SSL_VERIFYPEER, true);
+            curl_setopt($curl, CURLOPT_SSL_VERIFYHOST, 2);
+            curl_setopt($curl, CURLOPT_CAINFO, $this->_config->caFile());
+        }
+
+        if (!empty($requestBody)) {
+            curl_setopt($curl, CURLOPT_POSTFIELDS, $requestBody);
+        }
+
+        curl_setopt($curl, CURLOPT_RETURNTRANSFER, true);
+        $response = curl_exec($curl);
+        $httpStatus = curl_getinfo($curl, CURLINFO_HTTP_CODE);
+        curl_close($curl);
+        if ($this->_config->sslOn()) {
+            if ($httpStatus == 0) {
+                throw new Exception\SSLCertificate();
+            }
+        }
+        return array('status' => $httpStatus, 'body' => $response);
     }
 }