+++ src/Braintree/OAuthGateway.php
@@ -19,29 +19,70 @@ class OAuthGateway
     {
         $this->_gateway = $gateway;
         $this->_config = $gateway->config;
-        $this->_http = new HttpOAuth($gateway->config);
+        $this->_http = new Http($gateway->config);
+        $this->_http->useClientCredentials();
 
         $this->_config->assertHasClientCredentials();
     }
 
     public function createTokenFromCode($params)
     {
-        $params['grantType'] = 'authorization_code';
-        $response = $this->_http->post('/oauth/access_tokens', $params);
-        return $this->_verifyGatewayResponse($response);
+        $params['grantType'] = "authorization_code";
+        return $this->_createToken($params);
     }
 
     public function createTokenFromRefreshToken($params)
     {
-        $params['grantType'] = 'refresh_token';
+        $params['grantType'] = "refresh_token";
+        return $this->_createToken($params);
+    }
+
+    private function _createToken($params)
+    {
+        $params = array('credentials' => $params);
         $response = $this->_http->post('/oauth/access_tokens', $params);
         return $this->_verifyGatewayResponse($response);
     }
 
     private function _verifyGatewayResponse($response)
     {
-        $result = OAuthCredentials::factory($response);
-        $result->success = !isset($response['error']);
+        if (isset($response['credentials'])) {
+            $result =  new Result\Successful(
+                OAuthCredentials::factory($response['credentials'])
+            );
+            return $this->_mapSuccess($result);
+        } elseif (isset($response['apiErrorResponse'])) {
+            $result = new Result\Error($response['apiErrorResponse']);
+            return $this->_mapError($result);
+        } else {
+            throw new Exception\Unexpected(
+                "Expected credentials or apiErrorResponse"
+            );
+        }
+    }
+
+    public function _mapError($result)
+    {
+        $error = $result->errors->deepAll()[0];
+
+        if ($error->code == Error\Codes::OAUTH_INVALID_GRANT) {
+            $result->error = 'invalid_grant';
+        } elseif ($error->code == Error\Codes::OAUTH_INVALID_CREDENTIALS) {
+            $result->error = 'invalid_credentials';
+        } elseif ($error->code == Error\Codes::OAUTH_INVALID_SCOPE) {
+            $result->error = 'invalid_scope';
+        }
+        $result->errorDescription = explode(': ', $error->message)[1];
+        return $result;
+    }
+
+    public function _mapSuccess($result)
+    {
+        $credentials = $result->credentials;
+        $result->accessToken = $credentials->accessToken;
+        $result->refreshToken = $credentials->refreshToken;
+        $result->tokenType = $credentials->tokenType;
+        $result->expiresAt = $credentials->expiresAt;
         return $result;
     }
 
@@ -49,7 +90,8 @@ class OAuthGateway
     {
         $query = Util::camelCaseToDelimiterArray($params, '_');
         $query['client_id'] = $this->_config->getClientId();
-        $url = $this->_config->baseUrl().'/oauth/connect?'.http_build_query($query);
+        $queryString = preg_replace('/\%5B\d+\%5D/', '%5B%5D', http_build_query($query));
+        $url = $this->_config->baseUrl() . '/oauth/connect?' . $queryString;
 
         return $this->signUrl($url);
     }
@@ -57,6 +99,6 @@ class OAuthGateway
     private function signUrl($url)
     {
         $key = hash('sha256', $this->_config->getClientSecret(), true);
-        return $url.'&signature='.hash_hmac('sha256', $url, $key).'&algorithm=SHA256';
+        return $url . '&signature=' . hash_hmac('sha256', $url, $key) . '&algorithm=SHA256';
     }
 }