+++ src/Braintree/WebhookNotification.php
@@ -27,6 +27,8 @@ class WebhookNotification extends Braintree
         if (preg_match("/[^A-Za-z0-9+=\/\n]/", $payload) === 1) {
             throw new Exception\InvalidSignature('payload contains illegal characters');
         }
+
+        Configuration::assertGlobalHasAccessTokenOrKeys();
         self::_validateSignature($signature, $payload);
 
         $xml = base64_decode($payload);
@@ -40,6 +42,8 @@ class WebhookNotification extends Braintree
         if (!preg_match('/^[a-f0-9]{20,32}$/', $challenge)) {
             throw new Exception\InvalidChallenge("challenge contains non-hex characters");
         }
+
+        Configuration::assertGlobalHasAccessTokenOrKeys();
 
         $publicKey = Configuration::publicKey();
         $digest = Digest::hexDigestSha1(Configuration::privateKey(), $challenge);